+++ app/src/androidTest/java/com/oyelekeokiki/ui/BaseCartImplModelTest.kt
@@ -1,4 +1,4 @@
-package com.oyelekeokiki
+package com.oyelekeokiki.ui
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.oyelekeokiki.networking.RemoteApi
