+++ app/src/androidTest/java/com/oyelekeokiki/ui/CartViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.oyelekeokiki
+package com.oyelekeokiki.ui
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
